Modularize database schema tests
Separate migration, encrypted replica, cache, attachment, search, and entity schema coverage into focused modules while leaving production schema behavior unchanged.
